feat: rename MatrixClient folder
